Include IBEA code inside bluepyopt Compiler now became dependency New subpackage 'deapext' Backward compatibility is maintained for the moment
