+++ bluepyopt/__init__.py
@@ -19,15 +19,25 @@ Copyright (c) 2016, EPFL/Blue Brain Project
  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 """
 
+# pylint: disable=W0611
+
 from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
 
 from .api import *  # NOQA
 import bluepyopt.optimisations
+import bluepyopt.deapext.optimisations
+
+# Add some backward compatibility for the time when DEAPoptimisation not in
+# deapext yet
+# TODO deprecate this
+bluepyopt.optimisations.DEAPOptimisation = \
+    bluepyopt.deapext.optimisations.DEAPOptimisation
+
 import bluepyopt.evaluators
 import bluepyopt.objectives
-import bluepyopt.parameters
+import bluepyopt.parameters  # NOQA
 
 # TODO let objects read / write themselves using json
 # TODO create 'Variables' class
