+++ bluepyopt/deapext/tools/__init__.py
@@ -1,0 +1,3 @@
+"""Init"""
+
+from .selIBEA import *  # NOQA
