+++ requirements.txt
@@ -1,2 +1,1 @@
 -e .
-http://github.com/BlueBrain/deap/tarball/bbp-specific
